# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is a line bot for party usage",
        "title": "party line bot",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/images/list": {
            "get": {
                "description": "Get a list of images with optional filtering by name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "List images",
                "operationId": "list-images",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image name to filter by",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.imageResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/images/mark": {
            "post": {
                "description": "Mark an image with the specified ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "Mark an image",
                "operationId": "mark-image",
                "parameters": [
                    {
                        "description": "JSON payload with image ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.markRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.imageResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "controller.markRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}